Refactor synthetic data out of fixtures
